requirements.txt
Flask==1.1.1
jinja2<3.1
itsdangerous<2.1
werkzeug<2.1
urllib3<2
python-dotenv
requests